pyproject.toml
[project]
name = "pinecone-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-community>=0.3.20",
    "langchain-openai>=0.3.11",
    "langchain-pinecone>=0.2.3",
    "pinecone-notebooks>=0.1.1",
    "pypdf>=5.4.0",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "codespell>=2.4.1",
    "pre-commit>=4.2.0",
    "ruff>=0.11.12",
]
